UserAgentCore config no longer dependent on UA
