+++ src/Core/user-agents/user-agent-server.ts
@@ -249,7 +249,7 @@ export class UserAgentServer implements IncomingRequest {
       options.toTag = this.toTag;
     }
     options.userAgent = options.userAgent || this.core.configuration.userAgentHeaderFieldValue;
-    options.supported = options.supported || this.core.configuration.supportedResponseOptions;
+    options.supported = options.supported || this.core.configuration.supportedOptionTagsResponse;
     const response = constructOutgoingResponse(this.message, options);
     this.transaction.receiveResponse(options.statusCode, response.message);
     return response;
